backend/requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Image processing
pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.3

# ML & embeddings
open-clip-torch==2.24.0
torch==2.1.2
torchvision==0.16.2

# Vector search
faiss-cpu==1.7.4

# HTTP requests
requests==2.31.0
httpx==0.26.0

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
tqdm==4.66.1
python-dotenv==1.0.0
